appengine_config: logging: manually find and inject Google Cloud trace id
makes log lines attach to their HTTP request and show grouped in cloud console log browser. google-cloud-logging used to do this automatically for webapp2, based on the X-Cloud-Trace-Context header, but it dropped webapp2 as of google-cloud-logging 2.0.

background:
https://github.com/googleapis/python-logging/issues/110#issuecomment-745534629
https://github.com/googleapis/python-logging/issues/149#issuecomment-782693201

also note that AppEngineHandler is evidently deprecated, so I may need to port that whole class into webutil eventually. :(
https://github.com/googleapis/python-logging/issues/202